# ConcurrentQueueSegment allows spinning threads to sleep.