--- a/src/libraries/System.Private.CoreLib/src/System/Collections/Concurrent/ConcurrentQueueSegment.cs
+++ b/src/libraries/System.Private.CoreLib/src/System/Collections/Concurrent/ConcurrentQueueSegment.cs
@@ -37,6 +37,9 @@ namespace System.Collections.Concurrent
         internal ConcurrentQueueSegment<T>? _nextSegment; // SOS's ThreadPool command depends on this name
 #pragma warning restore 0649
 
+        /// <summary>Threshold to spin before allowing threads to sleep when there is contention.</summary>
+        internal const int _sleep1Threshold = 8;
+
         /// <summary>Creates the segment.</summary>
         /// <param name="boundedLength">
         /// The maximum number of elements the segment can contain.  Must be a power of 2.
@@ -183,7 +186,7 @@ namespace System.Collections.Concurrent
                 }
 
                 // Lost a race. Spin a bit, then try again.
-                spinner.SpinOnce(sleep1Threshold: -1);
+                spinner.SpinOnce(sleep1Threshold: _sleep1Threshold);
             }
         }
 
@@ -244,7 +247,7 @@ namespace System.Collections.Concurrent
                 }
 
                 // Lost a race. Spin a bit, then try again.
-                spinner.SpinOnce(sleep1Threshold: -1);
+                spinner.SpinOnce(sleep1Threshold: _sleep1Threshold);
             }
         }
 
@@ -301,7 +304,7 @@ namespace System.Collections.Concurrent
                 }
 
                 // Lost a race. Spin a bit, then try again.
-                spinner.SpinOnce(sleep1Threshold: -1);
+                spinner.SpinOnce(sleep1Threshold: _sleep1Threshold);
             }
         }