Proposal to fix [this issue](https://github.com/dotnet/runtime/issues/44077). The `SpinWait` instances in `ConcurrentQueueSegment` does not allow enqueuers/dequeuers to sleep when there is contention, causing a lot of time spent busy waiting.

This change would essentially undo [this merge](https://github.com/dotnet/coreclr/pull/21722) and reimplements a threshold value for the spinners. Originally I used the `Thread.OptimalMaxSpinWaitsPerSpinIteration` as the threshold value, which improved throughput on my EPYC machine significantly, but I've found that using a constant value of `8` results in very similar performance, which `Thread.OptimalMaxSpinWaitsPerSpinIteration` usually evaluates to anyways.

I ran both the Microbenchmark and TechEmpower benchmarks to evaluate the impact of the change. Here are the results for my Ryzen and Skylake machines:

### Microbenchmarks

Base - Current `master` branch
Diff - This change

**Ryzen**

```
summary:
better: 2, geomean: 5.838
worse: 1, geomean: 1.020
total diff: 3

| Slower                                                              | diff/base | Base Median (ns) | Diff Median (ns) | Modality|
| ------------------------------------------------------------------- | ---------:| ----------------:| ----------------:| --------:|
| System.Collections.IterateForEach<Int32>.ConcurrentQueue(Size: 512) |      1.02 |          4394.97 |          4483.57 |         |

| Faster                                                                           | base/diff | Base Median (ns) | Diff Median (ns) | Modality|
| -------------------------------------------------------------------------------- | ---------:| ----------------:| ----------------:| --------:|
| System.Collections.Concurrent.AddRemoveFromSameThreads<String>.ConcurrentQueue(S |      6.82 |     676441708.50 |      99189562.00 |         |
| System.Collections.Concurrent.AddRemoveFromSameThreads<Int32>.ConcurrentQueue(Si |      5.00 |     608823740.50 |     121816517.50 |         |
```

**Skylake**

```
summary:
better: 2, geomean: 2.179
worse: 1, geomean: 1.045
total diff: 3

| Slower                                                               | diff/base | Base Median (ns) | Diff Median (ns) | Modality|
| -------------------------------------------------------------------- | ---------:| ----------------:| ----------------:| --------:|
| System.Collections.IterateForEach<String>.ConcurrentQueue(Size: 512) |      1.05 |          5536.67 |          5787.78 |         |

| Faster                                                                           | base/diff | Base Median (ns) | Diff Median (ns) | Modality|
| -------------------------------------------------------------------------------- | ---------:| ----------------:| ----------------:| --------:|
| System.Collections.Concurrent.AddRemoveFromSameThreads<Int32>.ConcurrentQueue(Si |      2.23 |     188272565.50 |      84290311.00 |         |
| System.Collections.Concurrent.AddRemoveFromSameThreads<String>.ConcurrentQueue(S |      2.13 |     191395707.50 |      90050370.50 |         |
```


### TechEmpower

Default - Current `master` branch
`x.x.8` - Sleep threshold value of `8`

#### Plaintext

**Ryzen**

```
| application           | zen2-lin.plaintext.default | zen2-lin.plaintext.8 |        |
| --------------------- | -------------------------- | -------------------- | ------ |
| CPU Usage (%)         |                         51 |                   50 | -1.96% |
| Raw CPU Usage (%)     |                     606.53 |               601.99 | -0.75% |
| Working Set (MB)      |                        109 |                  112 | +2.75% |
| Build Time (ms)       |                      3,790 |                3,812 | +0.58% |
| Start Time (ms)       |                        218 |                  207 | -5.05% |
| Published Size (KB)   |                     94,124 |               94,124 |  0.00% |
| .NET Core SDK Version |        5.0.100-rtm.20509.5 |  5.0.100-rtm.20509.5 |        |


| load                   | zen2-lin.plaintext.default | zen2-lin.plaintext.8 |         |
| ---------------------- | -------------------------- | -------------------- | ------- |
| CPU Usage (%)          |                         51 |                   50 |  -1.96% |
| Raw CPU Usage (%)      |                     607.74 |               597.67 |  -1.66% |
| Working Set (MB)       |                         51 |                   52 |  +1.96% |
| Build Time (ms)        |                      3,673 |                3,829 |  +4.25% |
| Start Time (ms)        |                          0 |                    0 |         |
| Published Size (KB)    |                     76,404 |               76,404 |   0.00% |
| .NET Core SDK Version  |                    3.1.403 |              3.1.403 |         |
| Requests/sec           |                  5,837,436 |            5,840,076 |  +0.05% |
| Requests               |                 88,120,000 |           88,179,648 |  +0.07% |
| Mean latency (ms)      |                       6.72 |                 5.84 | -13.10% |
| Max latency (ms)       |                     354.35 |                83.95 | -76.31% |
| Bad responses          |                          0 |                    0 |         |
| Socket errors          |                          0 |                    0 |         |
| Read throughput (MB/s) |                     701.44 |               701.76 |  +0.05% |
| Latency 50th (ms)      |                       3.31 |                 3.32 |  +0.30% |
| Latency 75th (ms)      |                       8.46 |                 8.34 |  -1.42% |
| Latency 90th (ms)      |                      15.52 |                14.74 |  -5.03% |
| Latency 99th (ms)      |                      48.95 |                30.92 | -36.83% |
```

**Skylake**

```
| application         | skylake-lin.plaintext.default | skylake-lin.plaintext.8 |         |
| ------------------- | ----------------------------- | ----------------------- | ------- |
| CPU Usage (%)       |                            52 |                      52 |   0.00% |
| Raw CPU Usage (%)   |                        413.76 |                  414.74 |  +0.24% |
| Working Set (MB)    |                            93 |                     113 | +21.51% |
| Build Time (ms)     |                         2,816 |                   2,843 |  +0.96% |
| Start Time (ms)     |                           250 |                     247 |  -1.20% |
| Published Size (KB) |                        94,187 |                  94,187 |   0.00% |


| load                   | skylake-lin.plaintext.default | skylake-lin.plaintext.8 |         |
| ---------------------- | ----------------------------- | ----------------------- | ------- |
| CPU Usage (%)          |                            49 |                      49 |   0.00% |
| Raw CPU Usage (%)      |                        393.67 |                  391.69 |  -0.50% |
| Working Set (MB)       |                            51 |                      51 |   0.00% |
| Build Time (ms)        |                         3,186 |                   3,185 |  -0.03% |
| Start Time (ms)        |                             0 |                       0 |         |
| Published Size (KB)    |                        76,404 |                  76,404 |   0.00% |
| First Request (ms)     |                            66 |                      62 |  -6.06% |
| Requests/sec           |                     2,735,736 |               2,721,966 |  -0.50% |
| Requests               |                    41,299,696 |              41,096,432 |  -0.49% |
| Mean latency (ms)      |                          9.21 |                   10.83 | +17.59% |
| Max latency (ms)       |                        220.10 |                  293.38 | +33.29% |
| Bad responses          |                             0 |                       0 |         |
| Socket errors          |                             0 |                       0 |         |
| Read throughput (MB/s) |                        328.73 |                  327.08 |  -0.50% |
| Latency 50th (ms)      |                          5.22 |                    5.84 | +11.88% |
| Latency 75th (ms)      |                         12.70 |                   13.91 |  +9.53% |
| Latency 90th (ms)      |                         21.52 |                   25.28 | +17.47% |
| Latency 99th (ms)      |                         53.23 |                   75.93 | +42.65% |
```

#### Fortunes

**Ryzen**

```
| db                | zen2-lin.fortunes.default | zen2-lin.fortunes.8 |         |
| ----------------- | ------------------------- | ------------------- | ------- |
| CPU Usage (%)     |                        26 |                  27 |  +3.85% |
| Raw CPU Usage (%) |                    314.99 |              319.26 |  +1.36% |
| Working Set (MB)  |                       103 |                  98 |  -4.85% |
| Build Time (ms)   |                     4,794 |               4,305 | -10.20% |
| Start Time (ms)   |                     2,297 |               1,655 | -27.95% |


| application           | zen2-lin.fortunes.default | zen2-lin.fortunes.8 |        |
| --------------------- | ------------------------- | ------------------- | ------ |
| CPU Usage (%)         |                        51 |                  51 |  0.00% |
| Raw CPU Usage (%)     |                    609.16 |              607.95 | -0.20% |
| Working Set (MB)      |                       346 |                 348 | +0.58% |
| Build Time (ms)       |                     3,732 |               3,752 | +0.54% |
| Start Time (ms)       |                    11,869 |              11,879 | +0.08% |
| Published Size (KB)   |                    94,126 |              94,126 |  0.00% |
| .NET Core SDK Version |       5.0.100-rtm.20509.5 | 5.0.100-rtm.20509.5 |        |


| load                   | zen2-lin.fortunes.default | zen2-lin.fortunes.8 |        |
| ---------------------- | ------------------------- | ------------------- | ------ |
| CPU Usage (%)          |                        16 |                  16 |  0.00% |
| Raw CPU Usage (%)      |                    190.99 |              193.67 | +1.41% |
| Working Set (MB)       |                        51 |                  52 | +1.96% |
| Build Time (ms)        |                     3,783 |               3,683 | -2.64% |
| Start Time (ms)        |                         0 |                   0 |        |
| Published Size (KB)    |                    76,404 |              76,404 |  0.00% |
| .NET Core SDK Version  |                   3.1.403 |             3.1.403 |        |
| First Request (ms)     |                       132 |                 124 | -6.06% |
| Requests/sec           |                   145,481 |             146,284 | +0.55% |
| Requests               |                 2,193,818 |           2,207,425 | +0.62% |
| Mean latency (ms)      |                      3.55 |                3.52 | -0.85% |
| Max latency (ms)       |                     27.79 |               26.16 | -5.87% |
| Bad responses          |                         0 |                   0 |        |
| Socket errors          |                         0 |                   0 |        |
| Read throughput (MB/s) |                    188.83 |              189.87 | +0.55% |
| Latency 50th (ms)      |                      3.34 |                3.32 | -0.60% |
| Latency 75th (ms)      |                      4.13 |                4.10 | -0.73% |
| Latency 90th (ms)      |                      5.06 |                5.03 | -0.59% |
| Latency 99th (ms)      |                      8.28 |                8.10 | -2.17% |
```

**Skylake**

```
| db                | skylake-lin.fortunes.default | skylake-lin.fortunes.8 |        |
| ----------------- | ---------------------------- | ---------------------- | ------ |
| CPU Usage (%)     |                           30 |                     30 |  0.00% |
| Raw CPU Usage (%) |                       237.31 |                 238.85 | +0.65% |
| Working Set (MB)  |                           83 |                     78 | -6.02% |
| Build Time (ms)   |                        1,672 |                  1,688 | +0.96% |
| Start Time (ms)   |                          325 |                    302 | -7.08% |


| application         | skylake-lin.fortunes.default | skylake-lin.fortunes.8 |        |
| ------------------- | ---------------------------- | ---------------------- | ------ |
| CPU Usage (%)       |                           48 |                     48 |  0.00% |
| Raw CPU Usage (%)   |                       382.07 |                 381.20 | -0.23% |
| Working Set (MB)    |                          186 |                    180 | -3.23% |
| Build Time (ms)     |                        2,807 |                  2,822 | +0.53% |
| Start Time (ms)     |                       11,321 |                 11,310 | -0.10% |
| Published Size (KB) |                       94,189 |                 94,189 |  0.00% |


| load                   | skylake-lin.fortunes.default | skylake-lin.fortunes.8 |        |
| ---------------------- | ---------------------------- | ---------------------- | ------ |
| CPU Usage (%)          |                           18 |                     19 | +5.56% |
| Raw CPU Usage (%)      |                       146.53 |                 148.22 | +1.15% |
| Working Set (MB)       |                           51 |                     51 |  0.00% |
| Build Time (ms)        |                        3,163 |                  3,182 | +0.60% |
| Start Time (ms)        |                            0 |                      0 |        |
| Published Size (KB)    |                       76,404 |                 76,404 |  0.00% |
| First Request (ms)     |                           75 |                     76 | +1.33% |
| Requests/sec           |                       79,003 |                 78,921 | -0.10% |
| Requests               |                    1,192,522 |              1,191,434 | -0.09% |
| Mean latency (ms)      |                         6.49 |                   6.51 | +0.31% |
| Max latency (ms)       |                        43.32 |                  46.11 | +6.44% |
| Bad responses          |                            0 |                      0 |        |
| Socket errors          |                            0 |                      0 |        |
| Read throughput (MB/s) |                       102.54 |                 102.44 | -0.10% |
| Latency 50th (ms)      |                         6.17 |                   6.15 | -0.32% |
| Latency 75th (ms)      |                         7.55 |                   7.58 | +0.40% |
| Latency 90th (ms)      |                         9.09 |                   9.18 | +0.99% |
| Latency 99th (ms)      |                        13.71 |                  14.16 | +3.28% |
```

#### Json

**Ryzen**

```
| application           | zen2-lin.json.default | zen2-lin.json.8     |        |
| --------------------- | --------------------- | ------------------- | ------ |
| CPU Usage (%)         |                    59 |                  58 | -1.69% |
| Raw CPU Usage (%)     |                703.46 |              691.37 | -1.72% |
| Working Set (MB)      |                   267 |                 273 | +2.25% |
| Build Time (ms)       |                 3,742 |               3,761 | +0.51% |
| Start Time (ms)       |                   208 |                 208 |  0.00% |
| Published Size (KB)   |                94,124 |              94,124 |  0.00% |
| .NET Core SDK Version |   5.0.100-rtm.20509.5 | 5.0.100-rtm.20509.5 |        |


| load                   | zen2-lin.json.default | zen2-lin.json.8 |         |
| ---------------------- | --------------------- | --------------- | ------- |
| CPU Usage (%)          |                    42 |              42 |   0.00% |
| Raw CPU Usage (%)      |                504.94 |          501.70 |  -0.64% |
| Working Set (MB)       |                    51 |              51 |   0.00% |
| Build Time (ms)        |                 3,788 |           3,729 |  -1.56% |
| Start Time (ms)        |                     0 |               0 |         |
| Published Size (KB)    |                76,404 |          76,404 |   0.00% |
| .NET Core SDK Version  |               3.1.403 |         3.1.403 |         |
| Requests/sec           |               558,740 |         549,529 |  -1.65% |
| Requests               |             8,436,538 |       8,296,629 |  -1.66% |
| Mean latency (ms)      |                  4.63 |            2.96 | -36.07% |
| Max latency (ms)       |                412.12 |           91.91 | -77.70% |
| Bad responses          |                     0 |               0 |         |
| Socket errors          |                     0 |               0 |         |
| Read throughput (MB/s) |                 77.80 |           76.51 |  -1.66% |
| Latency 50th (ms)      |                  0.54 |            0.51 |  -4.81% |
| Latency 75th (ms)      |                  3.30 |            3.29 |  -0.30% |
| Latency 90th (ms)      |                 10.52 |            9.00 | -14.45% |
| Latency 99th (ms)      |                 71.37 |           25.65 | -64.06% |
```

**Skylake**

```
| application         | skylake-lin.json.default | skylake-lin.json.8 |        |
| ------------------- | ------------------------ | ------------------ | ------ |
| CPU Usage (%)       |                       58 |                 59 | +1.72% |
| Raw CPU Usage (%)   |                   467.18 |             468.79 | +0.34% |
| Working Set (MB)    |                      166 |                167 | +0.60% |
| Build Time (ms)     |                    2,820 |              2,775 | -1.60% |
| Start Time (ms)     |                      247 |                247 |  0.00% |
| Published Size (KB) |                   94,187 |             94,187 |  0.00% |


| load                   | skylake-lin.json.default | skylake-lin.json.8 |         |
| ---------------------- | ------------------------ | ------------------ | ------- |
| CPU Usage (%)          |                       51 |                 44 | -13.73% |
| Raw CPU Usage (%)      |                   410.26 |             353.84 | -13.75% |
| Working Set (MB)       |                       50 |                 51 |  +2.00% |
| Build Time (ms)        |                    3,314 |              3,147 |  -5.04% |
| Start Time (ms)        |                        0 |                  0 |         |
| Published Size (KB)    |                   76,404 |             76,404 |   0.00% |
| First Request (ms)     |                       60 |                 60 |   0.00% |
| Requests/sec           |                  243,422 |            250,537 |  +2.92% |
| Requests               |                3,671,284 |          3,779,742 |  +2.95% |
| Mean latency (ms)      |                    51.96 |              10.66 | -79.48% |
| Max latency (ms)       |                 1,990.00 |             550.29 | -72.35% |
| Bad responses          |                        0 |                  0 |         |
| Socket errors          |                       74 |                  0 |         |
| Read throughput (MB/s) |                    33.89 |              34.88 |  +2.92% |
| Latency 50th (ms)      |                     1.14 |               1.20 |  +5.26% |
| Latency 75th (ms)      |                    17.00 |               7.48 | -56.00% |
| Latency 90th (ms)      |                   151.25 |              22.24 | -85.30% |
| Latency 99th (ms)      |                   777.46 |             163.01 | -79.03% |
```

…