@stephentoub 2020-11-05
> Originally I used the Thread.OptimalMaxSpinWaitsPerSpinIteration as the threshold value, which improved throughput on my EPYC machine significantly, but I've found that using a constant value of 8 results in very similar performance, which Thread.OptimalMaxSpinWaitsPerSpinIteration usually evaluates to anyways.

What about just using `SpinWait()` with no argument? If it's in the same ballpark, my preference would be to not introduce another magic number. Or maybe subsequently even tweak the number or uses internally.